+++ lib/i18n/strings.g.dart
@@ -4,9 +4,9 @@
 /// To regenerate, run: `dart run slang`
 ///
 /// Locales: 1
-/// Strings: 733
+/// Strings: 734
 ///
-/// Built on 2026-05-30 at 16:29 UTC
+/// Built on 2026-05-31 at 09:47 UTC
 
 // coverage:ignore-file
 // ignore_for_file: type=lint, unused_import